Make -Wconversion and -Werror happy

+++ apps/client.cpp
@@ -101,7 +101,7 @@ int main(int argc, char* argv[]) {
                 VLOG(0) << "-- OnConnected";
                 X509* server_cert = SSL_get_peer_certificate(socketConnection->getSSL());
                 if (server_cert != nullptr) {
-                    int verifyErr = SSL_get_verify_result(socketConnection->getSSL());
+                    long verifyErr = SSL_get_verify_result(socketConnection->getSSL());
 
                     VLOG(0) << "     Server certificate: " + std::string(X509_verify_cert_error_string(verifyErr));
